+++ src/components/MyPage/MyResume/styles.js
@@ -1,10 +1,26 @@
 import styled from "styled-components";
 
+export const Container = styled.div`
+  width: 898px;
+  border: 2px solid;
+  border-radius: 12px;
+  border-color: #c9c9c9;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+`;
+
+export const Title = styled.label`
+  margin-top: 35px;
+  margin-left: 50px;
+  font-size: 24px;
+  font-weight: 700;
+`;
+
 export const ResumeButton = styled.button`
-  position: absolute;
-  left: 50%;
-  top: 113px;
-  transform: translateX(-50%);
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 32px;
   width: 750px;
   height: 40px;
   border: none;
@@ -20,10 +36,10 @@ export const ResumeButton = styled.button`
 `;
 
 export const CardContainter = styled.div`
-  position: relative;
-  left: 50%;
-  top: 179px;
-  transform: translateX(-50%);
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 27px;
+  margin-bottom: 50px;
   width: 767px;
   height: 650px;
   display: flex;
