+++ src/components/MyPage/MyTeamPage/styled.js
@@ -1,15 +1,23 @@
 import styled from "styled-components";
 
+export const Container = styled.div`
+  width: 898px;
+  border: 2px solid;
+  border-radius: 12px;
+  border-color: #c9c9c9;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+`;
+
 export const Title = styled.label`
-  position: absolute;
-  left: 36px;
-  top: 122px;
+  margin-left: 36px;
+  margin-top: 30px;
   font-size: 24px;
   font-weight: 700;
 `;
 
 export const Profile = styled.button`
-  position: absolute;
   left: 64px;
   top: 219px;
   width: 250px;
@@ -20,13 +28,11 @@ export const Profile = styled.button`
 `;
 
 export const Tag = styled.label`
-  position: absolute;
   font-size: 18px;
   font-weight: 600;
 `;
 
 export const Input = styled.input`
-  position: absolute;
   width: 332px;
   height: 36px;
   border: 1.9px solid;
@@ -35,7 +41,6 @@ export const Input = styled.input`
 `;
 
 export const Sel = styled.select`
-  position: absolute;
   width: 332px;
   height: 36px;
   border: 1.9px solid;
@@ -44,9 +49,8 @@ export const Sel = styled.select`
 `;
 
 export const Explanation = styled.textarea`
-  position: absolute;
   width: 496px;
-  height: 249px;
+  height: 229px;
   border: 1.9px solid;
   border-radius: 6px;
   border-color: #c9c9c9;
@@ -57,9 +61,6 @@ export const Explanation = styled.textarea`
 `;
 
 export const SaveButton = styled.button`
-  position: absolute;
-  right: 87px;
-  top: 752px;
   width: 95px;
   height: 35px;
   font-size: 18px;
